reorganized and added readme
