+++ .github/workflows/shared-stack-deploy.yaml
@@ -46,7 +46,6 @@ jobs:
           cache: "pip"
           cache-dependency-path: |
             aws_resources/requirements.txt
-            aws_resources/schema_initializer/requirements.txt
             aws_resources/map_population/requirements.txt
 
       - name: Install Python dependencies
@@ -54,7 +53,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-          pip install --upgrade --target schema_initializer -r schema_initializer/requirements.txt
           pip install -r map_population/requirements.txt
 
       - name: Install jq
@@ -123,7 +121,7 @@ jobs:
           echo "db_secret_arn=$DB_SECRET_ARN" >> $GITHUB_OUTPUT
 
       - name: Initialize RDS schema
-        working-directory: aws_resources/schema_initializer
+        working-directory: aws_resources/map_population
         env:
           DB_SECRET_ARN: ${{ steps.shared-outputs.outputs.db_secret_arn }}
           AWS_REGION: us-east-1
