+++ aws_resources/schema_initializer/populate_rds.py
@@ -1,184 +1,12 @@
-'''
-Initializes the database schema for indoor navigation.
-Triggers on any update or creation of the lambda during deployment.
-'''
-
-import os
-import json
-import pg8000
-import boto3
-import logging
 import sys
+from pathlib import Path
+import runpy
 
-logger = logging.getLogger()
-logger.setLevel(logging.INFO)
-
-DESTRUCTIVE_RESET_FLAG = "SCHEMA_INIT_ALLOW_DESTRUCTIVE_RESET"
-
-def get_db_secret():
-    """Retrieves database credentials from AWS Secrets Manager."""
-    secret_arn = os.environ.get('DB_SECRET_ARN')
-    if not secret_arn:
-        logger.error("DB_SECRET_ARN environment variable is not set.")
-        exit(1)
-        
-    client = boto3.client('secretsmanager')
-    response = client.get_secret_value(SecretId=secret_arn)
-    return json.loads(response['SecretString'])
-
-
-def ensure_destructive_reset_is_explicit():
-    """
-    Guardrail: schema init currently drops and recreates map tables.
-    Require explicit operator opt-in to avoid accidental data loss.
-    """
-    allow_reset = os.environ.get(DESTRUCTIVE_RESET_FLAG, "").strip().lower()
-    if allow_reset not in ("1", "true", "yes"):
-        logger.error(
-            "Refusing destructive schema reset. Set %s=true to continue.",
-            DESTRUCTIVE_RESET_FLAG,
-        )
-        logger.error(
-            "This tool owns DDL only. Run map_population/cli.py populate separately for map seed data."
-        )
-        sys.exit(1)
 
 def main():
-    print("Starting Schema Initialization...")
-    ensure_destructive_reset_is_explicit()
-    
-    # 1. Get Credentials
-    creds = get_db_secret()
-    
-    # 2. Connect to the Database
-    conn = None
-    try:
-        conn = pg8000.connect(
-            user=creds['username'],
-            password=creds['password'],
-            host=creds['host'],
-            port=int(creds['port']),
-            database=creds['dbname'] # This usually defaults to 'postgres' or the name you set in CDK
-        )
-        cursor = conn.cursor()
-        logger.info("Connected to Database.")
-        
-        # 3. Define the SQL Schema
-        # Note: We use DOUBLE PRECISION for coordinates to ensure accuracy for navigation.
-        # We use VARCHAR with CHECK constraints instead of ENUMs for easier updates later.
-        
-        # ... inside handler() ...
-        
-        # Cleanup existing tables if they exist.
-        # NOTE: This is intentionally destructive while migration tooling is pending.
-        cleanup_commands = [
-            "DROP TABLE IF EXISTS Rooms CASCADE;", 
-            "DROP TABLE IF EXISTS Landmarks CASCADE;", # Drops the conflict source
-            "DROP TABLE IF EXISTS MapEdges CASCADE;",
-            "DROP TABLE IF EXISTS MapNodes CASCADE;",
-            "DROP TABLE IF EXISTS Floors CASCADE;",
-            "DROP TABLE IF EXISTS Buildings CASCADE;"
-        ]
-
-        # Create tables
-        create_commands = [
-            # Buildings
-            """
-            CREATE TABLE Buildings (
-                BuildingID VARCHAR(50) PRIMARY KEY,
-                Name VARCHAR(255) NOT NULL,
-                GPS_Lat DOUBLE PRECISION,   -- Correct Name
-                GPS_Long DOUBLE PRECISION   -- Correct Name
-            );
-            """,
-            
-            # Floors
-            """
-            CREATE TABLE Floors (
-                FloorID SERIAL PRIMARY KEY,
-                BuildingID VARCHAR(50) REFERENCES Buildings(BuildingID) ON DELETE CASCADE,
-                FloorNumber INT NOT NULL,
-                MapImageURL TEXT,
-                MapScaleRatio DOUBLE PRECISION,
-                UNIQUE(BuildingID, FloorNumber)
-            );
-            """,
-            
-            # MapNodes
-            """
-            CREATE TABLE MapNodes (
-                -- Use the human-readable node ID string as the primary key
-                NodeIDString VARCHAR(255) PRIMARY KEY,
-                -- Keep an auto-incrementing integer for internal use and foreign keys
-                NodeID SERIAL UNIQUE,
-                FloorID INT REFERENCES Floors(FloorID) ON DELETE CASCADE,
-                BuildingID VARCHAR(50) REFERENCES Buildings(BuildingID),
-                CoordinateX INT NOT NULL,
-                CoordinateY INT NOT NULL,
-                NodeType VARCHAR(32) CHECK (NodeType IN ('Intersection', 'Corner', 'Elevator', 'Stairwell', 'Door', 'HallwayPoint')),
-                NodeMeta JSONB
-            );
-            """,
-            
-            # MapEdges
-            """
-            CREATE TABLE MapEdges (
-                EdgeID SERIAL PRIMARY KEY,
-                FloorID INT REFERENCES Floors(FloorID) ON DELETE CASCADE,
-                StartNodeID VARCHAR(255) REFERENCES MapNodes(NodeIDString) ON DELETE CASCADE,
-                EndNodeID VARCHAR(255) REFERENCES MapNodes(NodeIDString) ON DELETE CASCADE,
-                DistanceMeters DOUBLE PRECISION NOT NULL,
-                Bearing DOUBLE PRECISION,
-                IsBidirectional BOOLEAN DEFAULT TRUE
-            );
-            """,
-            
-            # Landmarks
-            """
-            CREATE TABLE Landmarks (
-                LandmarkID SERIAL PRIMARY KEY, -- Correct Name
-                FloorID INT REFERENCES Floors(FloorID) ON DELETE CASCADE,
-                Name VARCHAR(50) NOT NULL, -- e.g. "Room 205" or "Men's Restroom"
-                NearestNodeID VARCHAR(255) REFERENCES MapNodes(NodeIDString),
-                DistanceToNode DOUBLE PRECISION,
-                -- Legacy/optional free-form field; no cardinal CHECK constraint.
-                BearingFromNode VARCHAR(32),
-                MapCoordinateX INT,
-                MapCoordinateY INT
-            );
-            """
-        ]
-        
-        # Indexes for performance optimization
-        index_commands = [
-            "CREATE INDEX idx_mapnodes_floor ON MapNodes(FloorID);",
-            "CREATE INDEX idx_mapedges_floor ON MapEdges(FloorID);",
-            "CREATE INDEX idx_landmarks_floor ON Landmarks(FloorID);",
-            "CREATE INDEX idx_landmarks_name ON Landmarks(Name);"
-        ]
-
-        # EXECUTION LOOP
-        for sql in cleanup_commands:
-            cursor.execute(sql)
-            
-        for sql in create_commands:
-            cursor.execute(sql)
-            
-        for sql in index_commands:
-            cursor.execute(sql)
-            
-        conn.commit()
-        print("Schema successfully initialized.")
-        
-    except Exception as e:
-            logger.error(f"❌ Error initializing schema: {e}")
-            print(f"❌ Error initializing schema: {e}")
-            if conn:
-                conn.rollback()
-            exit(1)
-    finally:
-        if conn:
-            conn.close()
+    """Compatibility wrapper. Canonical script moved to map_population/populate_rds.py."""
+    target = Path(__file__).resolve().parents[1] / "map_population" / "populate_rds.py"
+    runpy.run_path(str(target), run_name="__main__")
 
 if __name__ == "__main__":
     main()
