+++ docs/reference.md
@@ -32,7 +32,7 @@
 | WebSocket API | Bidirectional live navigation/object-detection message transport | AWS API Gateway (WebSocket) | [openapi.yaml](./openapi.yaml), [cdk_stack.py](../aws_resources/cdk/cdk_stack.py) | WebSocket Lambda handlers, DynamoDB session state |
 | Backend Lambda handlers | Request handling for auth, navigation, and object detection | AWS Lambda  | [AuthHandler.kt](../aws_resources/backend/src/main/kotlin/com/handlers/AuthHandler.kt), [StaticNavigationHandler.kt](../aws_resources/backend/src/main/kotlin/com/handlers/StaticNavigationHandler.kt), [LiveNavigationHandler.kt](../aws_resources/backend/src/main/kotlin/com/handlers/LiveNavigationHandler.kt), [ObjectDetectionHandler.kt](../aws_resources/backend/src/main/kotlin/com/handlers/ObjectDetectionHandler.kt) | API Gateway, Cognito, RDS, DynamoDB, HTTP inference endpoint |
 | Identity service | User registration/authentication and token issuance | AWS Cognito User Pool | [cdk_stack.py](../aws_resources/cdk/cdk_stack.py), [openapi.yaml](./openapi.yaml) | Auth Lambda handler |
-| Relational map datastore | Persistent indoor map graph and landmark metadata (one shared Postgres instance, `StrideSharedStack`) | Amazon RDS (PostgreSQL) | [shared_stack.py](../aws_resources/cdk/shared_stack.py), [cdk_stack.py](../aws_resources/cdk/cdk_stack.py), [populate_rds.py](../aws_resources/schema_initializer/populate_rds.py) | Backend map/navigation services |
+| Relational map datastore | Persistent indoor map graph and landmark metadata (one shared Postgres instance, `StrideSharedStack`) | Amazon RDS (PostgreSQL) | [shared_stack.py](../aws_resources/cdk/shared_stack.py), [cdk_stack.py](../aws_resources/cdk/cdk_stack.py), [populate_rds.py](../aws_resources/map_population/populate_rds.py) | Backend map/navigation services |
 | Runtime state/config store | Session state and object-class height config | Amazon DynamoDB | [cdk_stack.py](../aws_resources/cdk/cdk_stack.py), [populate_obj_ddb.py](../aws_resources/object_config_seed/populate_obj_ddb.py) | Live/object-detection handlers |
 | Inference service | Computer vision inference used by object detection handlers | HTTP service (`/invocations`) | [HttpInferenceClient.kt](../aws_resources/backend/src/main/kotlin/com/services/HttpInferenceClient.kt), [ObjectDetectionHandler.kt](../aws_resources/backend/src/main/kotlin/com/handlers/ObjectDetectionHandler.kt), [inference_server](../inference_server/README.md) | Object detection and live navigation handlers |
 
@@ -91,11 +91,11 @@ API Gateway selects the Lambda integration using **`$request.body.action`** (see
 
 | Table | Primary Key | Key Relationships | Important Columns | Used By | Source |
 | --- | --- | --- | --- | --- | --- |
-| `Buildings` | `BuildingID` | Parent of `Floors.BuildingID` | `BuildingName`, `BuildingAddress` | Map data loading and navigation context | [populate_rds.py](../aws_resources/schema_initializer/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
-| `Floors` | `FloorID` | `BuildingID -> Buildings.BuildingID` | `FloorName`, `FloorNumber`, `ImageURL`, `Scale` | Floor-level map and search context | [populate_rds.py](../aws_resources/schema_initializer/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
-| `MapNodes` | `NodeIDString` | `FloorID -> Floors.FloorID`, `BuildingID -> Buildings.BuildingID` | `CoordinateX`, `CoordinateY`, `NodeType` | Pathfinding graph nodes and landmark nearest-node linkage | [populate_rds.py](../aws_resources/schema_initializer/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
-| `MapEdges` | (composite edge by row) | `StartNodeID/EndNodeID -> MapNodes.NodeIDString`, `FloorID -> Floors.FloorID` | `Distance`, `Direction` | Pathfinding graph edges for navigation | [populate_rds.py](../aws_resources/schema_initializer/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
-| `Landmarks` | `LandmarkID` | `FloorID -> Floors.FloorID`, `NearestNodeID -> MapNodes.NodeIDString` | `Name`, `LandmarkType`, `CoordinateX`, `CoordinateY` | Search results and destination mapping | [populate_rds.py](../aws_resources/schema_initializer/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
+| `Buildings` | `BuildingID` | Parent of `Floors.BuildingID` | `BuildingName`, `BuildingAddress` | Map data loading and navigation context | [populate_rds.py](../aws_resources/map_population/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
+| `Floors` | `FloorID` | `BuildingID -> Buildings.BuildingID` | `FloorName`, `FloorNumber`, `ImageURL`, `Scale` | Floor-level map and search context | [populate_rds.py](../aws_resources/map_population/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
+| `MapNodes` | `NodeIDString` | `FloorID -> Floors.FloorID`, `BuildingID -> Buildings.BuildingID` | `CoordinateX`, `CoordinateY`, `NodeType` | Pathfinding graph nodes and landmark nearest-node linkage | [populate_rds.py](../aws_resources/map_population/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
+| `MapEdges` | (composite edge by row) | `StartNodeID/EndNodeID -> MapNodes.NodeIDString`, `FloorID -> Floors.FloorID` | `Distance`, `Direction` | Pathfinding graph edges for navigation | [populate_rds.py](../aws_resources/map_population/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
+| `Landmarks` | `LandmarkID` | `FloorID -> Floors.FloorID`, `NearestNodeID -> MapNodes.NodeIDString` | `Name`, `LandmarkType`, `CoordinateX`, `CoordinateY` | Search results and destination mapping | [populate_rds.py](../aws_resources/map_population/populate_rds.py), [DATABASE_SCHEMA_REPORT.md](./DATABASE_SCHEMA_REPORT.md) |
 
 ### DynamoDB
